test: add test cases for if-elseif-else statement
